week2-job-scorer/functions/requirements.txt
azure-functions
azure-cosmos>=4.7.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
openai>=1.42.0
pydantic>=2.8.0
requests>=2.31.0
jinja2>=3.1.0
